requirements.txt
requests
python-dotenv
pandas
gspread
google-auth
